NAFI/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
defusedxml==0.7.1
Django==5.0.6
django-allauth==0.63.3
django-crispy-forms==2.1
django-tinymce==4.0.0
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
oauthlib==3.2.2
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
requests-oauthlib==2.0.0
setuptools==70.1.0
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.1